backend/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
python-dotenv
httpx
pydantic[email]